% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metR_ArrowGrob.R
\name{arrowGrob}
\alias{arrowGrob}
\title{Create a grob representing an arrow}
\usage{
arrowGrob(x, y, angle, length, pivot, default.units = "npc", ...)
}
\arguments{
\item{default.units}{}
}
\value{
a \pkg{grid} grob.
}
\description{
Function to create a grob representing an arrow.
}
\details{
Code stolen from from Elio Campitelli (https://github.com/eliocamp/metR) and adapted
by WTS.
}
