% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBoundingBoxToBBox.R
\name{convertBoundingBoxToBBox}
\alias{convertBoundingBoxToBBox}
\title{Convert a bounding box (or NULL) to a sf::bbox object}
\usage{
convertBoundingBoxToBBox(boundingbox)
}
\arguments{
\item{boundingbox}{- bounding box: a matrix (ala sp), a raster::extent object, a list, or NULL}
}
\value{
a sf::bbox object (numeric vector with xmin, ymin,xmax,ymax)
}
\description{
Function to convert a bounding box (or NULL) to a sf::bbox object.
}
\details{
boundingBox=NULL return a bbox for the earth in degrees.

 @export
}
