% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLayerFromShapefile.R
\name{createLayerFromShapefile}
\alias{createLayerFromShapefile}
\title{Create a spatial layer (an sf or sp object) from a shapefile}
\usage{
createLayerFromShapefile(file, strCRS = getCRS("WGS84"), as.sf = TRUE)
}
\arguments{
\item{file}{- shapefile to read}

\item{strCRS}{- string describing the output coordinate reference system (CRS) [or NULL to use the original]}

\item{as.sf}{- flag to create layer as simple features default is TRUE as of 2018.09.28)}
}
\value{
a spatial dataset consistent with the tmap package, either a
simple features dataset or an sp dataset depending on whether as.sf is TRUE or FALSE.
}
\description{
This function reads a shapefile and creates a spatial layer, either as a
simple features dataset (using package \code{sf}) or as an sp dataset (using package \code{sp}).
}
\details{
Uses \code{tmaptools::read_shape} to read the shapefile. If requested, uses \code{transformCRS} to
convert the output layer CRS to that represented by strCRS.
}
