% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRasterLayer.R
\name{createRasterLayer}
\alias{createRasterLayer}
\title{Create a spatial raster layer from a dataframe and a bounding box}
\usage{
createRasterLayer(dfr, col, bbox, nx = 500, ny = 500,
  strCRS = getCRS("WGS84"), linear = TRUE, extrap = FALSE,
  duplicate = "error")
}
\arguments{
\item{dfr}{- spatial dataframe with data to rasterize}

\item{col}{- name of column in dataframe to rasterize}

\item{bbox}{- tmap-style bounding box}

\item{nx}{- number of cells in y direction}

\item{ny}{- number of cells in y direction}

\item{strCRS}{- string representation of CRS in PROJ4 format (default yields long/lat in WGS84)}
}
\value{
- raster layer
}
\description{
This function creates a spatial raster layer from a dataframe and a bounding box
}
\details{
Uses \code{akima::interp()}, \code{sp::SpatialPixelsDataFrame()}, and \code{raster::raster()}.
}
