% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSF_points.R
\name{createSF_points}
\alias{createSF_points}
\title{Create an \pkg{sf} (simple features) dataset with point geometries}
\usage{
createSF_points(
  dfr,
  xCol = "longitude",
  yCol = "latitude",
  crs = get_crs(4326),
  wrapDateline = FALSE
)
}
\arguments{
\item{dfr}{\itemize{
\item dataframe with columns for x,y coordinates for points
}}

\item{xCol}{\itemize{
\item name of column with x coordinates (e.g., longitudes)
}}

\item{yCol}{\itemize{
\item name of column with y coordinates (e.g., latitudes)
}}

\item{crs}{\itemize{
\item a coordinate reference system object convertible to a \code{sf::crs} object using \code{\link{get_crs}}
}}

\item{wrapDateline}{\itemize{
\item flag (T/F) to use 0 to 360 rather than -180 to 180 range for longitudes
}}
}
\value{
an \pkg{sf} dataframe: dfr with "geom" column appended
}
\description{
Function to create an \pkg{sf} (simple features) dataset with point geometries.
}
\details{
Uses packages \code{tibble}, \code{sf}, \code{dplyr}.
}
