% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSFC_points.R
\name{createSFC_points}
\alias{createSFC_points}
\title{Create an sf (simple features) column with point geometries from x,y coordinate vectors}
\usage{
createSFC_points(x, y, crs = sf::NA_crs_)
}
\arguments{
\item{x}{- vector of x coordinates}

\item{y}{- vector of y coordinates}

\item{crs}{- coordinate reference system: EPSG code or character with proj4string}
}
\value{
an object of class sfc_POINT
}
\description{
Function to create an sf column (simple features) with point geometries (i.e., an sfc_POINT object) from x,y coordinate vectors.
}
\details{
Uses package \code{sf}.
}
