% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropFeaturesToBBox.R
\name{cropFeaturesToBBox}
\alias{cropFeaturesToBBox}
\title{Crop features (a \code{sf::sf} or \code{sf::sfc} object) to a bounding box}
\usage{
cropFeaturesToBBox(obj, bbx)
}
\arguments{
\item{obj}{- object of class "sf" or "sfc"}

\item{bbx}{- object that can be converted to a sf:bbox (see details)}
}
\value{
- object of same type as \code{obj}, but cropped to the bbox limits
}
\description{
Function to crop features (a \code{sf::sf} or \code{sf::sfc} object) to a bounding box.
}
\details{
\code{bbx} can be any object that can be converted to a \code{sf::bbox}
using \code{getBBox}.
}
\note{
The bounding box must be in the same coordinate system as \code{obj}. If the
coordinate reference system for \code{bbx} is undefined, it is assumed to be the same as
that for \code{obj}.
}
