% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropMapToBBox.R
\name{cropMapToBBox}
\alias{cropMapToBBox}
\title{Crop a map to a bounding box}
\usage{
cropMapToBBox(map, bbx)
}
\arguments{
\item{map}{- tmap object}

\item{bbx}{- object convertible to a \code{sf::bbox}}
}
\value{
tmap object with bounding box set to projected bbox
}
\description{
Function to crop a map to a bounding box
}
\details{
bbx replaces the bbox on the map object. Prior to replacement,
bbx is projected to the same coordinate reference system as the map.
}
