% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metR_ArrowGrob.R
\name{.unit_ifnot}
\alias{.unit_ifnot}
\title{Convert quantity to one with units (if unitless)}
\usage{
.unit_ifnot(x, unit)
}
\arguments{
\item{x}{: value to attach units}

\item{unit}{: units to attach}
}
\value{
\code{x} with units attached
}
\description{
Function to convert quantity to one with units (if unitless).
}
\details{
Uses \code{\link[grid:unit]{grid::unit()}} to attach units.
}
