% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGeometryColumns.R
\name{findGeometryColumns}
\alias{findGeometryColumns}
\title{Determine the names of the \pkg{sf} geometry column(s) in a dataframe}
\usage{
findGeometryColumns(sf_dfr, verbose = FALSE)
}
\arguments{
\item{sf_dfr}{- dataframe for which to determine the names of the geometry column(s)}

\item{verbose}{- flag to print diagnostic info}
}
\value{
vector of column names
}
\description{
Function to determine the names of the \pkg{sf} geometry column(s) in a dataframe.
}
\details{
The input dataframe does not have to be a simple features (\pkg{sf}) dataframe
(i.e., one with class = "sf") to contain \pkg{sf} geometry columns. For example, a
\pkg{sf} dataframe (class = "sf") with multiple geometry columns on which \code{\link[sf]{st_drop_geometry}} is ca
is called will have the "active" geometry column dropped and the resultant class will be
"tbl_df", but the other geometry columns (those that were not the active one) will not have been dropped.
}
