% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActiveGeometryColumnName.R
\name{getActiveGeometryColumnName}
\alias{getActiveGeometryColumnName}
\title{Determine the name of the active geometry column in an \pkg{sf} dataframe}
\usage{
getActiveGeometryColumnName(sf_dfr)
}
\arguments{
\item{sf_dfr}{\itemize{
\item \pkg{sf} dataframe for which to determine the name of the active geometry column
}}
}
\value{
vector of column names
}
\description{
Function to determine the name of the active geometry column in an \pkg{sf} dataframe.
}
\details{
none.
}
