% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCRS.R
\name{getCRS}
\alias{getCRS}
\title{Get a string representation for a familiar coordinate reference system}
\usage{
getCRS(name)
}
\arguments{
\item{name}{- familiar name of the CRS}
}
\value{
string representation of the CRS, as given by \code{tmaptools::get_proj4}
}
\description{
Function to get a string representation for a familiar coordinate reference system.
}
\details{
Uses \code{tmaptools::get_proj4} to get a string representation of a familiar CRS based on lookup table for
the EPSG code.
Currently handles "WGS84" (EPSG 4326), "NAD83" (EPSG 4269), and "AlaskaAlbers" (EPSG 3338).
}
