% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackagedLayer.R
\name{getPackagedLayer}
\alias{getPackagedLayer}
\title{Get a spatial layer stored with this package}
\usage{
getPackagedLayer(layerName, as.sf = TRUE)
}
\arguments{
\item{layerName}{- name of layer to pull out}

\item{as.sf}{- TRUE/FALSE to return dataset as an sf or sp type}
}
\value{
an sf or sp dataset
}
\description{
Function to get a spatial layer stored with this package.
}
\details{
Possible layers to pull out are:\cr
 "EBS_StationLabels"         ("ebsstationlabels.shp")\cr
 "EBS_SurveyBlocks"          ("NMFS_EBSSurveyBlocks.shp")\cr
 "EBS_SurveyStations"        ("NMFS_EBSSurveyStations.PointsLL.shp")\cr
 "EBS_SurveyStrata"          ("SurveyStrata_StandardArea.shp")\cr
 "NBS_SurveyBlocks"          ("northern_blocks.shp")\cr
 "NBS_SurveyStrata"          ("SurveyStrata_NBSArea.shp")\cr
 "NW_SurveyStrata"           ("SurveyStrata_NWArea.shp")\cr
 "BSAI_StatAreas"            ("NMFS_Statistical_Areas/BSAI.shp")\cr
 "ADFG_StatAreas"            ("ADFG_Statistical_Areas/ADFG_StatAreas_EBS.shp")\cr
 "Alaska"                    ("Land/Alaska.shp")\cr
 "ShelfBathymetry"           ("Bathymetry/ShelfBathymetry.shp")\cr
 "HCA_PribilofIslands"       ("Conservation_Areas/pribilof_hca.shp")\cr
}
