% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardBBox.R
\name{getStandardBBox}
\alias{getStandardBBox}
\title{Create a \code{sf::bbox} for a standard map area}
\usage{
getStandardBBox(area = c("EBS", "CGOA"))
}
\arguments{
\item{area}{\itemize{
\item area to return the bounding box for ("EBS","CGOA")
}}
}
\value{
\itemize{
\item bounding box in list format
}
}
\description{
This function \code{sf::bbox} for a standard map area.
}
\details{
The \code{sf::bbox} is returned in vector format, with values defining
xmin, ymin, xmax, ymax limits for the area.
}
