% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStandardBoundingBox.R
\name{getStandardBoundingBox}
\alias{getStandardBoundingBox}
\title{Create a bounding box to display a standard map area}
\usage{
getStandardBoundingBox(area = c("EBS", "CGOA"))
}
\arguments{
\item{area}{- area to return the bounding box for ("EBS","CGOA")}
}
\value{
- bounding box in list format
}
\description{
This function returns a bounding box to display a standard map area.
}
\details{
The bounding box is returned in list format, with coordinates defining the
bottomleft and topright boundaries for the area.
}
