% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFeatureLayers.R
\name{listFeatureLayers}
\alias{listFeatureLayers}
\title{List the (vector) feature layers in a dsn}
\usage{
listFeatureLayers(dsn)
}
\arguments{
\item{dsn}{\itemize{
\item data source name (interpretation varies by driver)
}}
}
\value{
a named list by layer with sub-lists as elements which have
information about the layer.
}
\description{
This function creates a list the (vector) feature layers in a dsn.
}
\details{
None.
}
