% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metR_ArrowGrob.R
\name{makeContent.arrow2}
\alias{makeContent.arrow2}
\title{Make content for an arrow}
\usage{
\method{makeContent}{arrow2}(x)
}
\arguments{
\item{x}{: a grid grob}
}
\value{
a grid grob or gTree
}
\description{
Function to make content for an arrow.
}
\details{
This function will be used by grid.draw when called on an arrowGrob. It
provides a specific implementation for the generic \code{\link[grid:makeContent]{grid::makeContent()}}.
}
