% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metR_ArrowGrob.R
\name{makeContent.vector}
\alias{makeContent.vector}
\title{Make content for an arrow}
\usage{
\method{makeContent}{vector}(x)
}
\arguments{
\item{x}{- a vectorGrob}
}
\value{
a grid grob or gTree
}
\description{
Function to make content for a vectorGrob object.
}
\details{
This function will be used by grid.draw when called on a vectorGrob. It
provides a specific implementation for the generic grid::makeContent(x).
}
