% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFeatureLayer.R
\name{readFeatureLayer}
\alias{readFeatureLayer}
\title{Create a sf dataframe from a (vector) feature layer}
\usage{
readFeatureLayer(dsn, layer, drivers = character(0), crs = NULL)
}
\arguments{
\item{dsn}{-data source name (interpretation varies by driver)}

\item{layer}{\itemize{
\item feature layer to read (interpretation varies by driver)
}}

\item{drivers}{\itemize{
\item RGDAL drivers to use to read feature class (limited set of short names of drivers to try; default is to try all)
}}

\item{crs}{\itemize{
\item a \code{sf::crs} object, EPSG code, or string describing the output coordinate reference system (or NULL to keep the original)
}}
}
\value{
a \code{sf::sf} data.frame.
}
\description{
This function reads a (vector) feature layer from a dsn and creates a
simple features dataframe (using package \code{sf}).
}
\details{
The function uses \code{sf::st_read} to read the feature layer from the dsn.
If \code{crs} is NULL, the coordinate reference system for the layer is unchanged.
to it.

If the layer did not have a coordinate reference system originally associated with it, then
\code{crs} is used to \strong{set} the \code{sf} object's coordinate reference system.

If the layer has a coordinate reference system associated with it,
\code{wtsGIS::transformCRS} is used to \strong{transform/project} the \code{sf} object
to that represented by \code{crs}
}
