% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRasterLayer.R
\name{readRasterLayer}
\alias{readRasterLayer}
\title{Create a stars raster from a raster layer}
\usage{
readRasterLayer(dsn, driver = character(0))
}
\arguments{
\item{dsn}{- data source name (interpretation varies by driver)}

\item{driver}{- RGDAL drivers to use to read feature class (limited set of short names of drivers to try; default is to try all)}
}
\value{
a \code{stars::stars} object.
}
\description{
This function reads a raster layer from a dsn and creates a
stars object (using package \code{stars}).
}
\details{
The function uses \code{stars::read_stars} to read the feature layer from the dsn.
If \code{crs} is NULL, the coordinate reference system for the layer is unchanged.
to it.

If the layer did not have a coordinate reference system originally associated with it, then
\code{crs} is used to **set** the \code{sf} object's coordinate reference system.

If the layere has a coordinate reference system associated with it,
\code{wtsGIS::transformCRS} is used to **transform/project** the \code{sf} object
to that represented by \code{crs}
}
