% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStarsLayer.R
\name{readStarsLayer}
\alias{readStarsLayer}
\title{Read a raster file into a \pkg{stars} object}
\usage{
readStarsLayer(dsns, driver = character(0), curvilinear = character(0))
}
\arguments{
\item{dsns}{- data source(s) name (interpretation varies by driver)}

\item{driver}{- (optional) RGDAL driver to use to read raster layer}

\item{curvilinear}{- (optional) length two character vector with names of sub-datasets holding longitude and latitude values for all raster cells}
}
\value{
a \pkg{stars} object.
}
\description{
This function reads a raster from a dsn and creates a
stars object (using \code{\link[stars]{read_stars}}).
}
\details{
The function uses \code{\link[stars]{read_stars}} to read the raster from the dsns.
}
