% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBoundingBoxOnMap.R
\name{setBoundingBoxOnMap}
\alias{setBoundingBoxOnMap}
\title{Set a bounding box on a map}
\usage{
setBoundingBoxOnMap(
  map,
  bbox,
  crs_bbox = tmaptools::get_proj4("longlat", output = "character")
)
}
\arguments{
\item{map}{- tmap object}

\item{bbox}{- bbox object}

\item{crs_bbox}{- coordinate reference system for bbox}
}
\value{
tmap object with bounding box set to projected bbox
}
\description{
Function to set a bounding box on a map
}
\details{
Requires package \code{tmap} and \code{tmaptools}.
}
