% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_longitude.R
\name{shift_longitude}
\alias{shift_longitude}
\title{Shift longitude coordinates of \pkg{sf} object to 0-360}
\usage{
shift_longitude(obj)
}
\arguments{
\item{obj}{- object to shift coordinates}
}
\value{
object of same class as obj, but with longitudes shifted
}
\description{
Function to shift longitude coordinates of \pkg{sf} object to 0-360.
}
\details{
This function expands the \code{\link[sf]{st_shift_longitude}} function
to work with \code{sf::bbox} objects as well as other \pkg{sf} objects.

If the coordinate reference system of the object is not lat/lon, no shift is made.
}
