% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.CreateBaseMap.R
\name{tmap.CreateBaseMap}
\alias{tmap.CreateBaseMap}
\title{Create a basemap layer for maps based on the tmap package}
\usage{
tmap.CreateBaseMap(layer.land = NULL, layer.bathym = NULL, gisDir = NULL,
  shapeFile.land = system.file("extdata/Shapefiles/Land/Alaska.shp", package =
  "wtsGIS"),
  shapeFile.bathymetry = system.file("extdata/Shapefiles/Bathymetry/ShelfBathymetry.shp",
  package = "wtsGIS"), strCRS = tmaptools::get_proj4("longlat"),
  boundingbox = list(bottomleft = list(lon = -179, lat = 54), topright =
  list(lon = -157, lat = 62.5)), colors.bathym = "darkblue",
  points.size = 0.01, verbose = TRUE)
}
\arguments{
\item{layer.land}{- a tmap layer representing land}

\item{layer.bathym}{- a tmap layer representing bathymetry}

\item{gisDir}{- path to top level folder for shapefiles}

\item{shapeFile.land}{- land shapefile (if layer.land is not provided)}

\item{shapeFile.bathymetry}{- bathymetry shapefile (if layer.bathym is not provided)}

\item{strCRS}{- string representation of CRS (default = WGS84) used for ALL shapefiles}

\item{boundingbox}{- a tmap-style bounding box}

\item{colors.bathym}{- color for the bathymetry}

\item{points.size}{- size of points, in map units}

\item{verbose}{- flag to print debugging info}
}
\value{
- basemap layer based on the tmap package
}
\description{
This function creates a basemap layer for maps based on the tmap package.
}
\details{
The basemap contains a land layer (polygons) and a bathymetry layer (lines).
Uses \code{wtsUtilities::tmap.CreateLayerFromShapefile}.
}
