% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.CreateLayerFromShapefile.R
\name{tmap.CreateLayerFromShapefile}
\alias{tmap.CreateLayerFromShapefile}
\title{Create a tmap layer from a shapefile}
\usage{
tmap.CreateLayerFromShapefile(file, strCRS = tmaptools::get_proj4("longlat"))
}
\arguments{
\item{file}{- shapefile to read}

\item{strCRS}{- string describing the final coordinate reference system (CRS)}
}
\value{
a map layer consistent with the tmap package
}
\description{
This function reads a shapefile and creates a tmap layer.
}
\details{
None.
}
