% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.CreateRasterLayer.R
\name{tmap.CreateRasterLayer}
\alias{tmap.CreateRasterLayer}
\title{Create a raster from a dataframe and a bounding box}
\usage{
tmap.CreateRasterLayer(dfr, col, bbox, nx = 500, ny = 500,
  strCRS = tmaptools::get_proj4("longlat", output = "character"),
  linear = TRUE, extrap = FALSE, duplicate = "error")
}
\arguments{
\item{dfr}{- spatial dataframe with data to rasterize}

\item{col}{- name of column in dataframe to rasterize}

\item{bbox}{- tmap-style bounding box}

\item{nx}{- number of cells in y direction}

\item{ny}{- number of cells in y direction}

\item{strCRS}{- string representation of CRS in PROJ4 format (default yields long/lat in WGS84)}
}
\value{
- raster layer
}
\description{
This function creates a raster from a dataframe and a bounding box
}
\details{
Uses akima::interp(), sp::SpatialPixelsDataFrame(), and raster::raster().
}
