% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.getPackagedLayer.R
\name{tmap.getPackagedLayer}
\alias{tmap.getPackagedLayer}
\title{Get a layer stored with the package}
\usage{
tmap.getPackagedLayer(layerName, as.sf = TRUE)
}
\arguments{
\item{layerName}{- name of layer to pull out}

\item{as.sf}{- TRUE/FALSE to return dataset as an sf or sp type}
}
\value{
an sf or sp dataset
}
\description{
Function to get a layer stored with this package.
}
\details{
Possible layers to pull out are:\cr
 "EBS_StationLabels"         ("ebsstationlabels.shp")
 "EBS_SurveyBlocks"          ("NMFS_EBSSurveyBlocks.shp")
 "EBS_SurveyStations"        ("NMFS_EBSSurveyStations.PointsLL.shp")
 "EBS_SurveyStrata"          ("SurveyStrata_StandardArea.shp")
 "NBS_SurveyBlocks"          ("northern_blocks.shp")
 "NBS_SurveyStrata"          ("SurveyStrata_NBSArea.shp")
 "NW_SurveyStrata"           ("SurveyStrata_NWArea.shp")
 "BSAI_StatAreas"            ("NMFS_Statistical_Areas/BSAI.shp")
 "ADFG_StatAreas"            ("ADFG_Statistical_Areas/ADFG_StatAreas_EBS.shp")
 "Alaska"                    ("Land/Alaska.shp")
 "ShelfBathymetry"           ("Bathymetry/ShelfBathymetry.shp")
 "HCA_PribilofIslands"       ("Conservation_Areas/pribilof_hca.shp")
}
