% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap.MergeDataframeWithLayer.R
\name{tmap.MergeDataframeWithLayer}
\alias{tmap.MergeDataframeWithLayer}
\title{Merge a dataframe with a geometry layer}
\usage{
tmap.MergeDataframeWithLayer(dfr, geoms, dataID = "STATION_ID",
  geomsID = "GIS_STATION", allData = FALSE, duplicateGeoms = TRUE)
}
\arguments{
\item{dfr}{- dataframe to merge}

\item{geoms}{- geometries to merge with}

\item{dataID}{- name of id column in dataframe to join on}

\item{geomsID}{- name of id column in geometry layer to join on}

\item{allData}{- flag to merge all rows from dataframe}

\item{duplicateGeoms}{- allow geoms to be duplicated (i.e., where multiple rows in dataframe maatch a single geom)}
}
\value{
- merged spatial dataframe (based on sf or sp classes, depending on class of geoms)
}
\description{
This function merges a dataset with a geometry layer
}
\details{
If geoms is an sf-like spatial dataset, the merge is conducted using \code{sf::left_join}; otherwise,
it is conducted using \code{sp:merge}.
}
