% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_CreateBasemap.R
\name{tmap_CreateBasemap}
\alias{tmap_CreateBasemap}
\title{Create a basemap layer for maps based on the tmap package}
\usage{
tmap_CreateBasemap(
  layer.land = getPackagedLayer("Alaska"),
  layer.bathym = getPackagedLayer("ShelfBathymetry"),
  final.crs = get_crs("WGS84"),
  bbox = getStandardBBox("EBS"),
  colors.bg = "white",
  colors.land = "grey85",
  colors.bathym = "darkblue",
  alpha.bathym = 1,
  points.size = 0.01
)
}
\arguments{
\item{layer.land}{\itemize{
\item a \code{sf::sf} object representing land
}}

\item{layer.bathym}{\itemize{
\item a \code{sf::sf} object representing bathymetry
}}

\item{final.crs}{\itemize{
\item representation of final (display) coordinate reference system (see details)
}}

\item{bbox}{\itemize{
\item a bounding box (see details)
}}

\item{colors.bg}{\itemize{
\item background color
}}

\item{colors.land}{\itemize{
\item color for land
}}

\item{colors.bathym}{\itemize{
\item color for the bathymetry
}}

\item{alpha.bathym}{\itemize{
\item transparency for the bathymetry
}}

\item{points.size}{\itemize{
\item size of points, in map units
}}
}
\value{
\itemize{
\item basemap layer based on the tmap package
}
}
\description{
This function creates a basemap layer for maps based on the tmap package.
}
\details{
The basemap contains a land layer (polygons) and a bathymetry layer (lines).
Uses \code{\link[=getPackagedLayer]{getPackagedLayer()}} or \code{\link[=readShapefile]{readShapefile()}} to create the
land and bathymetric layers. If \code{bbox} is NULL, then the
bounding box for the land layer is used as the bounding box for the basemap.

The final coordinate reference system (\code{final.crs}) can be any object that
can be converted to a \code{\link[sf:st_crs]{sf::st_crs()}} object using \code{\link[=get_crs]{get_crs()}}.

The bounding box (\code{bbox}) can be any object that can be converted
to a \code{sf::bbox} using \code{\link[=getBBox]{getBBox()}}.
}
