% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_CropMapToBBox.R
\name{tmap_CropMapToBBox}
\alias{tmap_CropMapToBBox}
\title{Crop a \pkg{tmap}-style map to a bounding box}
\usage{
tmap_CropMapToBBox(map, bbx)
}
\arguments{
\item{map}{- \pkg{tmap}-style map object}

\item{bbx}{- object convertible to a \code{sf::bbox}}
}
\value{
\pkg{tmap}-style map object with bounding box set to projected bbox
}
\description{
Function to crop a \pkg{tmap}-style map to a bounding box
}
\details{
bbx replaces the bbox on the map object. Prior to replacement,
bbx is projected to the same coordinate reference system as the map.
}
