% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformBBox.R
\name{transformBBox}
\alias{transformBBox}
\title{Transform (or assign) a bounding box's coordinate reference system}
\usage{
transformBBox(bbx, crs)
}
\arguments{
\item{bbx}{- object convertible to a \code{sf::bbox}}

\item{crs}{- the output coordinate reference system (see @details)}
}
\value{
the transformed bounding box in the specifed crs
}
\description{
Function to transform (or assign) a bounding box's coordinate reference system.
}
\details{
\code{crs} should be an object convertible to a \code{sf::crs} by
\code{get_crs}. If the crs of the bbox is NA, then \code{crs} is assigned to
the bbox, but no coordinate transformation takes place.
}
