% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCRS.R
\name{transformCRS}
\alias{transformCRS}
\title{Transform a spatial dataset to a different coordinate reference system}
\usage{
transformCRS(layer, strCRS)
}
\arguments{
\item{layer}{- spatial data layer to transform}

\item{strCRS}{- string representation of final CRS}
}
\value{
spatial layer of same type as input (sf or sp)
}
\description{
Function to transform a spatial dataset to a different coordinate reference system.
}
\details{
Uses \code{tmaptools::set_projection}
}
