% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionBBox.R
\name{unionBBox}
\alias{unionBBox}
\title{Union two bounding box objects}
\usage{
unionBBox(bbx1 = NULL, bbx2 = NULL)
}
\arguments{
\item{bbx1}{- original bounding box}

\item{bbx2}{- updating bounding box}
}
\value{
resulting bounding box as an sf::bbox object
}
\description{
Function to union two bounding boxes
}
\details{
Returns a sf::bbox covering the union of the individual bounding boxes.
If not sf::bbox's, the bounding boxes are converted to sf::bbox objects, then unioned.
}
