% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metR_ArrowGrob.R
\name{vectorGrob}
\alias{vectorGrob}
\title{Create a grob defining a geometric vector}
\usage{
vectorGrob(
  x,
  y,
  dx,
  dy,
  length,
  preserve.dir,
  default.units = "npc",
  pivot,
  ...
)
}
\arguments{
\item{x}{: x-axis location of base (passed on to \code{\link[grid:grid.grob]{grid::grob()}})}

\item{y}{: y-axis location of base (passed on to \code{\link[grid:grid.grob]{grid::grob()}})}

\item{dx}{: x-axis offset to tip of vector}

\item{dy}{: y-axis offset to tip of vector}

\item{length}{: passed on to \code{\link[grid:grid.grob]{grid::grob()}}}

\item{preserve.dir}{: passed on to \code{\link[grid:grid.grob]{grid::grob()}}}

\item{default.units}{: units to attach to x, dx, y, or dy (if not already attached )}

\item{pivot}{: passed on to \code{\link[grid:grid.grob]{grid::grob()}}}

\item{...}{: passed on to \code{\link[grid:grid.grob]{grid::grob()}}}
}
\value{
grob describing a vector
}
\description{
Function to create a grob defining a geometric vector.
}
\details{
Basically a pass-through for \code{\link[grid:grid.grob]{grid::grob()}}
}
