% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTableToCSV.R
\name{writeTableToCSV}
\alias{writeTableToCSV}
\title{Write a table (without geometry) as a csv file}
\usage{
writeTableToCSV(tbl, file = "table.csv")
}
\arguments{
\item{tbl}{\itemize{
\item table (sf, tibble, or dataframe object)
}}

\item{file}{\itemize{
\item file name
}}
}
\value{
invisible NULL
}
\description{
Function to write a table (without geometry) as a csv file.
}
\details{
Requires \code{sf} package. If \code{tbl} is a
\code{sf} object (i.e., with geometry information), the
geometry information is dropped.
}
