% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTableToShapefile.R
\name{writeTableToShapefile}
\alias{writeTableToShapefile}
\title{Write a table (with geometry information) as a shapefile}
\usage{
writeTableToShapefile(tbl, file = "table.shp")
}
\arguments{
\item{tbl}{- table object with geometry information (i.e., it inherits from \code{sf})}

\item{file}{- file name}
}
\value{
none
}
\description{
Function to write a table (with geometry information) as a shapefile.
}
\details{
Wrapper for \code{tmaptools::write_shape(tbl,file=file)}.
}
