% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToGISDataset.R
\name{writeToGISDataset}
\alias{writeToGISDataset}
\title{Write a geospatial object to a GIS dataset}
\usage{
writeToGISDataset(
  obj,
  driver = "ESRI Shapefile",
  dsn = ".",
  delete_dsn = FALSE,
  layer = NULL,
  delete_layer = TRUE
)
}
\arguments{
\item{obj}{\itemize{
\item a geospatial object to convert to GIS dataset
}}

\item{driver}{\itemize{
\item name of dataset driver (run \code{sf::st_drivers()} to see installed drivers )
}}

\item{dsn}{\itemize{
\item dsn for output (folder or file, depending on output)
}}

\item{delete_dsn}{\itemize{
\item flag to delete the dsn before writing
}}

\item{layer}{\itemize{
\item name of layer written to (assuming dsn has layers), or NULL
}}

\item{delete_layer}{\itemize{
\item flag to delete the layer before writing
}}
}
\value{
an invisible NULL.
}
\description{
Function to write a geospatial object (e.g., an sf dataset) to a GIS dataset.
}
\details{
Wrapper for functions \code{sf::st_write} for objects of class 'sf' or 'sfc' or
\code{rgdal::writeOGR} for objects of class Spatial.

#--former importFrom rgdal writeOGR
}
