% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToShapefile.R
\name{writeToShapefile}
\alias{writeToShapefile}
\title{Write an object with geometry information as a shapefile}
\usage{
writeToShapefile(obj, file = "shapefile")
}
\arguments{
\item{obj}{- object with geometry information}

\item{file}{- file name (default is "shapefile"; no extension is necessary)}
}
\value{
none
}
\description{
Function to write an object with geometry information (e.g. an sf dataframe) as a shapefile.
}
\details{
Wrapper for \code{writeToGISDataset} using "ESRI Shapefile" as the driver.
Note that column names are limited to 10 characters max.
}
