% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBaseTMap.R
\name{createBaseTMap}
\alias{createBaseTMap}
\title{Create a basemap layer for maps based on the tmap package}
\usage{
createBaseTMap(
  layer.land = getPackagedLayer("Alaska"),
  layer.bathym = getPackagedLayer("ShelfBathymetry"),
  crs.finl = get_crs("WGS84"),
  bbox = getStandardBBox("EBS"),
  colors.bg = "white",
  colors.land = "grey85",
  colors.bathym = "darkblue",
  alpha.bathym = 1,
  points.size = 0.01
)
}
\arguments{
\item{layer.land}{- a \code{sf::sf} object representing land}

\item{layer.bathym}{- a \code{sf::sf} object representing bathymetry}

\item{crs.finl}{- representation of final (display) coordinate reference system (see details)}

\item{bbox}{- a bounding box (see details)}

\item{colors.bg}{- background color}

\item{colors.land}{- color for land}

\item{colors.bathym}{- color for the bathymetry}

\item{alpha.bathym}{- transparency for the bathymetry}

\item{points.size}{- size of points, in map units}
}
\value{
- basemap layer based on the tmap package
}
\description{
This function creates a basemap layer for maps based on the tmap package.
}
\details{
The basemap contains a land layer (polygons) and a bathymetry layer (lines).
Uses \code{getPackagedLayer} or \code{wtsUtilities::readShapefile} to create the
land and bathymetric layers. If \code{bbox} is NULL, then the
bounding box for the land layer is used as the bounding box for the basemap.

The final coordinate reference system (\code{crs.finl}) can be any object that
can be converted to a \code{sf::crs} object using \code{get_crs}.

The bounding box (\code{bbox}) can be any object that can be converted
to a \code{sf::bbox} using \code{getBBox}.
}
