% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSF_points.R
\name{createSF_points}
\alias{createSF_points}
\title{Create an sf (simple features) dataset with point geometries}
\usage{
createSF_points(
  dfr,
  xCol = "longitude",
  yCol = "latitude",
  strCRS = get_proj4string("longlat")
)
}
\arguments{
\item{dfr}{- dataframe with columns for x,y coordinates for points}

\item{xCol}{- name of column with x coordinates (e.g., longitudes)}

\item{yCol}{- name of column with y coordinates (e.g., latitudes)}

\item{strCRS}{- proj4string representation of a coordinate reference system}
}
\value{
an sf dataframe with "geom" column appended to dfr
}
\description{
Function to create an sf (simple features) dataset with point geometries.
}
\details{
Uses packages \code{tibble}, \code{sf}, \code{dplyr}, \code{tmaptools}.
}
