% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSFC_points.R
\name{createSFC_points}
\alias{createSFC_points}
\title{Create an \pkg{sf} (simple features) column with point geometries from x,y coordinate vectors}
\usage{
createSFC_points(x, y, crs = sf::NA_crs_, wrapDateline = FALSE)
}
\arguments{
\item{x}{- vector of x coordinates}

\item{y}{- vector of y coordinates}

\item{crs}{- coordinate reference system (convertible to an sf crs by \code{\link{get_crs}})}

\item{wrapDateline}{- flag (T/F) to use 0 to 360 rather than -180 to 180 range for longitudes}
}
\value{
an object of class sfc_POINT
}
\description{
Function to create an \pkg{sf} (simple features) column with point geometries (i.e., an sfc_POINT object) from x,y coordinate vectors.
}
\details{
Uses package \pkg{sf}.
}
