% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackagedLayer.R
\name{getPackagedLayer}
\alias{getPackagedLayer}
\title{Get a spatial layer (as a \pkg{sf} dataframe) stored with this package}
\usage{
getPackagedLayer(layerName)
}
\arguments{
\item{layerName}{- name of layer to pull out}
}
\value{
a \pkg{sf} dataframe
}
\description{
Function to get a spatial layer (as a \pkg{sf} dataframe) stored with this package.
}
\details{
Possible layers to pull out are:
\itemize{
 \item{"EBS_StationLabels"         ("ebsstationlabels.shp")}
 \item{"EBS_SurveyBlocks"          ("NMFS_EBSSurveyBlocks.shp")}
 \item{"StandardCrabStations"      ("StandardCrabStations/StandardEBSSurveyStations_Crab.shp")}
 \item{"EBS_SurveyStations"        ("NMFS_EBSSurveyStations.PointsLL.shp")}
 \item{"EBS_SurveyStrata"          ("SurveyStrata_StandardArea.shp")}
 \item{"NBS_SurveyBlocks"          ("northern_blocks.shp")}
 \item{"NBS_SurveyStrata"          ("SurveyStrata_NBSArea.shp")}
 \item{"NW_SurveyStrata"           ("SurveyStrata_NWArea.shp")}
 \item{"BSAI_StatAreas"            ("NMFS_Statistical_Areas/BSAI.shp")}
 \item{"ADFG_StatAreas"            ("ADFG_Statistical_Areas/ADFG_StatAreas_EBS.shp")}
 \item{"Alaska"                    ("Land/Alaska.shp")}
 \item{"ShelfBathymetry"           ("Bathymetry/ShelfBathymetry.shp")}
 \item{"HCA_PribilofIslands"       ("Conservation_Areas/pribilof_hca.shp")}
}
}
