% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_CreateBasemapLayers.R
\name{gg_CreateBasemapLayers}
\alias{gg_CreateBasemapLayers}
\title{Create basemap layers for maps based on the \pkg{ggplot2} package}
\usage{
gg_CreateBasemapLayers(
  layer.land = getPackagedLayer("Alaska"),
  layer.bathym = getPackagedLayer("ShelfBathymetry"),
  final.crs = get_crs("WGS84"),
  bbox = getStandardBBox("EBS"),
  colors.bg = "white",
  colors.land = "grey85",
  colors.bathym = "darkblue",
  alpha.bathym = 1
)
}
\arguments{
\item{layer.land}{- a \code{sf::sf} object representing land}

\item{layer.bathym}{- a \code{sf::sf} object representing bathymetry}

\item{final.crs}{- object representing the final (display) coordinate reference system (see details)}

\item{bbox}{- a bounding box (see details; default is for the EBS)}

\item{colors.bg}{- background color (default is "white")}

\item{colors.land}{- color for land (default is "grey85")}

\item{colors.bathym}{- color for the bathymetry (default is "darkblue")}

\item{alpha.bathym}{- transparency for the bathymetry (default is 1)}
}
\value{
- basemap layer based on the \pkg{ggplot2} package
}
\description{
This function creates a basemap layer for maps based on the \pkg{ggplot2} package.
}
\details{
The basemap contains a land layer (polygons) and a bathymetry layer (lines).
Uses \code{\link{getPackagedLayer}} or \code{wtsUtilities::readShapefile} to create the
land and bathymetric layers. If \code{bbox} is NULL, then the
bounding box for the land layer is used as the bounding box for the basemap.

The final coordinate reference system (\code{final.crs}) can be any object that
can be converted to a \code{sf::crs} object using \code{\link{get_crs}}.

The bounding box (\code{bbox}) can be any object that can be converted
to a \code{sf::bbox} using \code{\link{getBBox}}.

layer.land and layer.bathym will first be transformed to the bounding box coordinates
and cropped to the bounding box. All three will then be converted to the final.crs
coordinate system.

If the bounding box is NULL, it will be created from the bounding box for
layer.land (if not NULL) or layer.bathym (if layer.land is NULL.

If final.crs is NULL, it will be set to the CRS for the bounding box.
}
