% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCRS.R
\name{transformCRS}
\alias{transformCRS}
\title{Transform a spatial dataset to a different coordinate reference system}
\usage{
transformCRS(layer, toCRS)
}
\arguments{
\item{layer}{\itemize{
\item spatial data layer to transform
}}

\item{toCRS}{\itemize{
\item representation of final coordinate reference system (see @details)
}}
}
\value{
spatial layer of same type as input (sf or sp)
}
\description{
Function to transform a spatial dataset to a different coordinate reference system.
}
\details{
If \code{layer} is an object of class sf, sfc, or sfg, then \code{toCRS}
should be convertible to an sf::crs object via \code{get_crs(toCRS)}.

if \code{layer} is a \code{sf::bbox} object, the bbox is transformed using
\code{transformBBox}.

#--If \code{layer} is an object of class Spatial, then \code{toCRS}
should be convertible to an sp::CRS object via \code{get_spCRS(toCRS)} .
}
