% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformCRS.R
\name{transformCRS}
\alias{transformCRS}
\title{Transform a spatial dataset to a different coordinate reference system}
\usage{
transformCRS(layer, toCRS)
}
\arguments{
\item{layer}{- spatial data layer to transform}

\item{toCRS}{- representation of final coordinate reference system (see @details)}
}
\value{
spatial layer of same type as input (sf or sp)
}
\description{
Function to transform a spatial dataset to a different coordinate reference system.
}
\details{
If \code{layer} is an object of class sf, sfc, or sfg, then \code{toCRS}
should be convertible to an sf::crs object via \code{get_crs(toCRS)}.

If \code{layer} is an object of class Spatial, then \code{toCRS}
should be convertible to an sp::CRS object via \code{get_spCRS(toCRS)} .
}
