% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionBBox.R
\name{unionBBox}
\alias{unionBBox}
\title{Union two bounding boxes}
\usage{
unionBBox(bbx1 = NULL, bbx2 = NULL)
}
\arguments{
\item{bbx1}{\itemize{
\item original bounding box
}}

\item{bbx2}{\itemize{
\item updating bounding box
}}
}
\value{
resulting bounding box as an sf::bbox object
}
\description{
Function to union two bounding boxes.
}
\details{
Returns a \code{sf::bbox} covering the union of the individual bounding boxes.
If not \code{sf::bbox}'s, the bounding boxes are converted to \code{sf::bbox} objects
using \code{getBBox}, then unioned.
}
