% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToGISDataset.R
\name{writeToGISDataset}
\alias{writeToGISDataset}
\title{Write a geospatial object to a GIS dataset}
\usage{
writeToGISDataset(
  obj,
  driver = "ESRI Shapefile",
  dsn = ".",
  delete_dsn = FALSE,
  layer = NULL,
  delete_layer = TRUE
)
}
\arguments{
\item{obj}{- a geospatial object to convert to GIS dataset}

\item{driver}{- name of dataset driver (run \code{sf::st_drivers()} to see installed drivers )}

\item{dsn}{- dsn for output (folder or file, depnding on output)}

\item{delete_dsn}{- flag to delete the dsn before writing}

\item{layer}{- name of layer written to (assuming dsn has layers), or NULL}

\item{delete_layer}{- flag to delete the layer before writing}
}
\value{
an invisible NULL.
}
\description{
Function to write a geospatial object (e.g., an sf dataset) to a GIS dataset.
}
\details{
Uses package \code{sf::st_write} or \code{rgdal::writeOGR}.
}
